move all socket io methods into seperate file
